+++ espr/src/parser/types/generalized.rs
@@ -12,9 +12,9 @@ pub fn named_types(input: &str) -> ParseResult<String> {
 /// 266 parameter_type = [generalized_types] | [named_types] | [simple_types] .
 pub fn parameter_type(input: &str) -> ParseResult<Type> {
     alt((
-        generalized_types,
         named_types.map(Type::Named),
         simple_types.map(Type::Simple),
+        generalized_types,
     ))
     .parse(input)
 }
@@ -150,7 +150,12 @@ mod tests {
         let (res, (set, _remarks)) = super::parameter_type("vim").finish().unwrap();
         dbg!(&set);
         assert_eq!(res, "");
-        assert_eq!(set, Type::Named("vim".to_string()),)
+        assert_eq!(set, Type::Named("vim".to_string()),);
+
+        let (res, (set, _remarks)) = super::parameter_type("generic_homhom").finish().unwrap();
+        dbg!(&set);
+        assert_eq!(res, "");
+        assert_eq!(set, Type::Named("generic_homhom".to_string()),);
     }
 
     #[test]